Extract config loading to a simple script
