+++ src/find-binary.js
